[HttpClient][Contracts] introduce component and related contracts
| Q             | A
| ------------- | ---
| Branch?       | master
| Bug fix?      | no
| New feature?  | yes
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | yes
| Fixed tickets | -
| License       | MIT
| Doc PR        | -

Sooner that later, we'll need to send HTTP calls to APIs,
e.g. Symfony Mailer or #27738 already need it.

Common existing HTTP clients for PHP rely on PSR-7,
which is orthogonal to the way Symfony is designed.

More reasons we need this in core are the package principles:
if we want to be able to keep our BC+deprecation promises, we
have to build on more stable and more abstract dependencies than
Symfony itself.

So here we are. This PR introduces a new `Http` namespace in contracts,
and a new `HttpClient` component.

Its surface is by design very simple to use, while still flexible
enough to cover more advanced use cases thanks to streaming+laziness.

```php
$client = new NativeHttpClient();

$responseContent = $client->get('https://google.com')->getContent();
```

Here is how to fetch some resources concurrently (this takes <3s with curl
and >12s with the native client):

```php
$client = new CurlHttpClient();
$pool = [];

for ($i = 0; $i < 379; ++$i) {
    $pool[] = $r = $client->get("https://http2.akamai.com/demo/tile-$i.png");
}

foreach ($r->complete(...$pool) as $r) {
    $r->getHeaders();
}
```

By default, clients follow redirections. When the redirection limit is
reached, or when a `4xx` or a `5xx` happens, an exception is thrown.

An array of options can be used when sending requests - see `HttpClientTrait`.

TODO:
 - [ ] validate the design
 - [ ] add tests
 - [ ] split in several PRs?

Implemented:
 - a simple contract for HTTP client
 - fopen()+curl-based clients
 - gzip compression enabled when possible
 - fetch multiple responses concurrently

Not implemented - help wanted:
 - handle standard proxy-related env vars when using NativeHttpClient;
   it might be implemented as a decorator.
 - handle redirections "manually"; it's required to handle proxy-related
   env vars properly; it might be implemented as a decorator.
 - a TraceableHttpClient and integration with the profiler
 - FrameworkBundle integration: autowireable alias + semantic configuration if needed
 - have NativeHttpClient honor socket.bindto set to a local Unix socket
 - ease authentication?